requirements.txt
pystray
watchdog
psutil
pywin32
plyer
pillow
